scraper/requirements.txt
requests
beautifulsoup4
pandas
pypdf
Pillow
pytesseract
